pyproject.toml
[project]
name = "axenta-geozones-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "drf-yasg>=1.21.15",
    "gunicorn>=25.3.0",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "ruff>=0.15.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
